# openapi.yaml
openapi: 3.0.3
info:
  title: Kasir API
  description: Backend API untuk aplikasi sistem kasir
  version: 1.0.0
  contact:
    name: Kasir Dev

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /:
    get:
      tags:
        - General
      summary: Root endpoint
      responses:
        "200":
          description: Pesan selamat datang
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Hello, Ini Backend Program Kasir!"

  /health:
    get:
      tags:
        - General
      summary: Health check
      responses:
        "200":
          description: Status layanan
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: "Service is running"

  /categories:
    get:
      tags:
        - Categories
      summary: Ambil semua kategori atau cari kategori berdasarkan nama
      description: |
        Mengambil daftar semua kategori. Jika parameter query `name` diberikan,
        akan melakukan pencarian kategori dengan nama yang mengandung kata kunci
        tersebut (case-insensitive).
        
        **Contoh request:**
        - Semua kategori: `GET /categories`
        - Cari berdasarkan nama: `GET /categories?name=Elec`
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Filter kategori berdasarkan nama (partial match, case-insensitive)
          example: Elec
      responses:
        "200":
          description: Daftar kategori (semua atau hasil pencarian)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
              examples:
                allCategories:
                  summary: Semua kategori
                  value:
                    - id: 1
                      name: "Electronics"
                      description: "Electronic devices and gadgets"
                    - id: 2
                      name: "Accessories"
                      description: "Related accessories and add-ons"
                searchResults:
                  summary: Hasil pencarian untuk "Elec"
                  value:
                    - id: 1
                      name: "Electronics"
                      description: "Electronic devices and gadgets"
                    - id: 3
                      name: "Electronic Accessories"
                      description: "Accessories for electronic devices"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - Categories
      summary: Buat kategori baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Kategori berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  /categories/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags:
        - Categories
      summary: Ambil kategori berdasarkan ID
      responses:
        "200":
          description: Detail kategori
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Categories
      summary: Update kategori
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Kategori berhasil diupdate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Categories
      summary: Hapus kategori
      responses:
        "200":
          description: Kategori berhasil dihapus
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Category deleted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/products:
    get:
      tags:
        - Products
      summary: Ambil semua produk atau cari produk berdasarkan nama
      description: |
        Mengambil daftar semua produk. Jika parameter query `name` diberikan,
        akan melakukan pencarian produk dengan nama yang mengandung kata kunci
        tersebut (case-insensitive).
        
        **Contoh request:**
        - Semua produk: `GET /api/products`
        - Cari berdasarkan nama: `GET /api/products?name=Lap`
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Filter produk berdasarkan nama (partial match, case-insensitive)
          example: Lap
      responses:
        "200":
          description: Daftar produk (semua atau hasil pencarian)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
              examples:
                allProducts:
                  summary: Semua produk
                  value:
                    - id: 1
                      name: "Laptop"
                      price: 999.99
                      stock: 10
                      category_id: 1
                      category_name: "Electronics"
                    - id: 2
                      name: "Smartphone"
                      price: 499.99
                      stock: 25
                      category_id: 1
                      category_name: "Electronics"
                searchResults:
                  summary: Hasil pencarian untuk "Lap"
                  value:
                    - id: 1
                      name: "Laptop"
                      price: 999.99
                      stock: 10
                      category_id: 1
                      category_name: "Electronics"
                    - id: 5
                      name: "Laptop Pro"
                      price: 1299.99
                      stock: 5
                      category_id: 1
                      category_name: "Electronics"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - Products
      summary: Buat produk baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Produk berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/products/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags:
        - Products
      summary: Ambil produk berdasarkan ID
      responses:
        "200":
          description: Detail produk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Products
      summary: Update produk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Produk berhasil diupdate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Products
      summary: Hapus produk
      responses:
        "200":
          description: Produk berhasil dihapus
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Product deleted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/checkout:
    post:
      tags:
        - Transactions
      summary: Checkout - Buat transaksi baru
      description: |
        Membuat transaksi baru dengan list produk yang dibeli.
        Sistem akan:
        - Validasi ketersediaan produk dan stok
        - Mengurangi stok produk otomatis
        - Menghitung total dan subtotal
        - Membuat record transaksi dan detail transaksi
        
        **Contoh request:**
        ```
        POST /api/checkout
        ```
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
            example:
              items:
                - product_id: 1
                  quantity: 2
                - product_id: 3
                  quantity: 1
      responses:
        "201":
          description: Transaksi berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
              example:
                id: 1
                total_amount: 2500
                created_at: "2026-02-08T08:54:56Z"
                details:
                  - id: 1
                    transaction_id: 1
                    product_id: 1
                    product_name: "Laptop"
                    quantity: 2
                    subtotal: 2000
                  - id: 2
                    transaction_id: 1
                    product_id: 3
                    product_name: "Mouse"
                    quantity: 1
                    subtotal: 500
        "400":
          description: Request tidak valid atau stok tidak cukup
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                emptyItems:
                  summary: Items kosong
                  value:
                    error: "Items cannot be empty"
                insufficientStock:
                  summary: Stok tidak cukup
                  value:
                    error: "insufficient stock for product Laptop (available: 5, requested: 10)"
                productNotFound:
                  summary: Produk tidak ditemukan
                  value:
                    error: "product with id 999 not found"

  /api/transactions:
    get:
      tags:
        - Transactions
      summary: Ambil semua transaksi
      description: |
        Mengambil daftar semua transaksi yang pernah dibuat,
        diurutkan berdasarkan tanggal terbaru.
        
        **Contoh request:**
        ```
        GET /api/transactions
        ```
      responses:
        "200":
          description: Daftar transaksi
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    total_amount:
                      type: integer
                    created_at:
                      type: string
                      format: date-time
              example:
                - id: 2
                  total_amount: 1500
                  created_at: "2026-02-08T08:50:00Z"
                - id: 1
                  total_amount: 2500
                  created_at: "2026-02-08T08:45:00Z"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/transactions/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags:
        - Transactions
      summary: Ambil detail transaksi
      description: |
        Mengambil detail lengkap transaksi termasuk semua item yang dibeli.
        
        **Contoh request:**
        ```
        GET /api/transactions/1
        ```
      responses:
        "200":
          description: Detail transaksi lengkap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
              example:
                id: 1
                total_amount: 2500
                created_at: "2026-02-08T08:45:00Z"
                details:
                  - id: 1
                    transaction_id: 1
                    product_id: 1
                    product_name: "Laptop"
                    quantity: 2
                    subtotal: 2000
                  - id: 2
                    transaction_id: 1
                    product_id: 3
                    product_name: "Mouse"
                    quantity: 1
                    subtotal: 500
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Transaksi tidak ditemukan
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "transaction not found"

  /api/report/hari-ini:
    get:
      tags:
        - Reports
      summary: Laporan transaksi hari ini
      description: |
        Mengambil ringkasan laporan transaksi untuk hari ini, meliputi:
        - Total revenue (pendapatan) hari ini
        - Total jumlah transaksi hari ini
        - Produk terlaris hari ini (berdasarkan quantity terjual)
        
        **Contoh request:**
        ```
        GET /api/report/hari-ini
        ```
      responses:
        "200":
          description: Laporan hari ini
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
              example:
                total_revenue: 45000
                total_transaksi: 5
                produk_terlaris:
                  nama: "Indomie Goreng"
                  qty_terjual: 12
        "500":
          $ref: "#/components/responses/InternalError"

  /api/report:
    get:
      tags:
        - Reports
      summary: Laporan transaksi berdasarkan range tanggal
      description: |
        Mengambil ringkasan laporan transaksi untuk periode tertentu, meliputi:
        - Total revenue (pendapatan) dalam periode
        - Total jumlah transaksi dalam periode
        - Produk terlaris dalam periode (berdasarkan quantity terjual)
        
        **Contoh request:**
        ```
        GET /api/report?start_date=2026-01-01&end_date=2026-02-01
        ```
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2026-01-01"
          description: Tanggal awal periode laporan (format YYYY-MM-DD)
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2026-02-01"
          description: Tanggal akhir periode laporan (format YYYY-MM-DD)
      responses:
        "200":
          description: Laporan periode yang ditentukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
              example:
                total_revenue: 150000
                total_transaksi: 15
                produk_terlaris:
                  nama: "Laptop"
                  qty_terjual: 25
        "400":
          description: Parameter tidak valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                missingStartDate:
                  summary: start_date tidak diberikan
                  value:
                    error: "start_date parameter is required (format: YYYY-MM-DD)"
                missingEndDate:
                  summary: end_date tidak diberikan
                  value:
                    error: "end_date parameter is required (format: YYYY-MM-DD)"
                invalidFormat:
                  summary: Format tanggal tidak valid
                  value:
                    error: "Invalid date format. Use YYYY-MM-DD"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
      example: 1

  schemas:
    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Electronics
        description:
          type: string
          example: Electronic devices and gadgets

    CategoryInput:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          example: Electronics
        description:
          type: string
          example: Electronic devices and gadgets

    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Laptop
        price:
          type: number
          format: double
          example: 999.99
        stock:
          type: integer
          example: 10
        category_id:
          type: integer
          example: 1
        category_name:
          type: string
          example: Electronics

    ProductInput:
      type: object
      required:
        - name
        - price
        - stock
        - category_id
      properties:
        name:
          type: string
          example: Laptop
        price:
          type: number
          format: double
          example: 999.99
        stock:
          type: integer
          example: 10
        category_id:
          type: integer
          example: 1

    Transaction:
      type: object
      properties:
        id:
          type: integer
          example: 1
        total_amount:
          type: integer
          description: "Total harga transaksi (dalam satuan terkecil, misal IDR)"
          example: 2500
        created_at:
          type: string
          format: date-time
          example: "2026-02-08T08:54:56Z"
        details:
          type: array
          items:
            $ref: "#/components/schemas/TransactionDetail"

    TransactionDetail:
      type: object
      properties:
        id:
          type: integer
          example: 1
        transaction_id:
          type: integer
          example: 1
        product_id:
          type: integer
          example: 1
        product_name:
          type: string
          description: Nama produk (optional, untuk kemudahan display)
          example: "Laptop"
        quantity:
          type: integer
          example: 2
        subtotal:
          type: integer
          description: "Harga total untuk item ini (price * quantity)"
          example: 2000

    CheckoutRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CheckoutItem"
          minItems: 1
          description: List produk yang akan dibeli
      example:
        items:
          - product_id: 1
            quantity: 2
          - product_id: 3
            quantity: 1

    CheckoutItem:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: integer
          description: ID produk yang akan dibeli
          example: 1
        quantity:
          type: integer
          description: Jumlah produk yang dibeli
          minimum: 1
          example: 2

    DailyReport:
      type: object
      properties:
        total_revenue:
          type: integer
          description: "Total pendapatan hari ini"
          example: 45000
        total_transaksi:
          type: integer
          description: "Total jumlah transaksi hari ini"
          example: 5
        produk_terlaris:
          $ref: "#/components/schemas/ProdukTerlaris"

    ProdukTerlaris:
      type: object
      properties:
        nama:
          type: string
          description: "Nama produk terlaris"
          example: "Indomie Goreng"
        qty_terjual:
          type: integer
          description: "Jumlah quantity terjual"
          example: 12

  responses:
    BadRequest:
      description: Request tidak valid
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Invalid product ID"
    NotFound:
      description: Data tidak ditemukan
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "produk tidak ditemukan"
    InternalError:
      description: Error internal server
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal server error"